requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==4.1.7
django-filter==22.1
django-psdb-engine==1.0.6
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
filelock==3.10.6
greenlet==2.0.2
identify==2.5.22
iniconfig==2.0.0
Markdown==3.4.3
msgpack==1.0.5
mypy==1.1.1
mypy-extensions==1.0.0
mysqlclient==2.1.1
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.1
Pygments==2.14.0
PyJWT==2.6.0
pynvim==0.4.3
pytest==7.2.2
python-decouple==3.8
pytz==2023.2
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2023.2.0.0
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
virtualenv==20.21.0
